Split BitmapFont into a separate module
